# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  content.UpdateShotRequest:
    properties:
      camera_movement:
        description: 运镜方式
        type: string
      duration:
        description: 时长（秒）
        type: number
      image_prompt:
        description: 图片提示词
        type: string
      narration:
        description: 解说内容
        type: string
      shot_id:
        description: 分镜头ID（必填）
        type: string
      video_prompt:
        description: 视频提示词
        type: string
    required:
    - shot_id
    type: object
  http.ErrorResponse:
    properties:
      code:
        description: 错误码（非0表示错误）
        type: integer
      detail:
        description: 错误详情（可选）
        type: string
      message:
        description: 错误消息
        type: string
    type: object
...
  resource.ErrorResponse:
    properties:
      code:
        description: 错误码（非0表示错误）
        type: integer
      detail:
        description: 错误详情（可选）
        type: string
      message:
        description: 错误消息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@lemon.ai
    name: API Support
  description: |-
    Lemon is an AI-powered API service built with Eino framework.
    It provides LLM chat, intelligent agent, and more AI capabilities.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Lemon API
  version: "1.0"
paths:
...
  /api/v1/resources/{resource_id}/download:
    get:
      consumes:
      - application/json
      description: 根据资源ID下载文件，返回文件流
      parameters:
      - description: 资源ID
        in: path
        name: resource_id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 文件流
          schema:
            type: file
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/resource.ErrorResponse'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/resource.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/resource.ErrorResponse'
      summary: 下载文件
      tags:
      - 资源管理
...
  /api/v1/shots:
    get:
      consumes:
      - application/json
      description: 根据章节ID和版本号获取镜头列表（按 index 排序）。镜头中包含解说内容。如果不传版本号，则获取最新版本。
      parameters:
      - description: 章节ID
        in: query
        name: chapter_id
        required: true
        type: string
      - description: 版本号（可选，不传则获取最新版本）
        in: query
        name: version
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: 获取镜头列表
      tags:
      - 内容管理
    put:
      consumes:
      - application/json
      description: 更新分镜头的脚本信息（解说、图片提示词、视频提示词、运镜方式、时长等）
      parameters:
      - description: 请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/content.UpdateShotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: 更新分镜头信息
      tags:
      - 内容管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"